+++ src/Graph_final.json
@@ -1,1 +1,362 @@
-{"nodes": {"2499": {"address": "Globoko", "lat": 45.9540673, "lon": 15.6339655, "uuid": "S2"}, "63502": {"address": "hrvaska5", "lat": 45.8936019, "lon": 15.6962282, "uuid": "H5"}, "47460": {"address": "hrvaska1", "lat": 45.839665, "lon": 15.6863823, "uuid": "H1"}, "17940": {"address": "Jesenice", "lat": 45.8595037, "lon": 15.6897957, "uuid": "S3"}, "18440": {"address": "hrvaska6", "lat": 45.8606795, "lon": 15.8031616, "uuid": "H6"}, "86774": {"address": "hrvaska2", "lat": 45.8031663, "lon": 15.7101405, "uuid": "H2"}, "25590": {"address": "Topliska cesta", "lat": 45.89337, "lon": 15.6214561, "uuid": "S4"}, "31310": {"address": "hrvaska3", "lat": 45.9171602, "lon": 15.7280016, "uuid": "H3"}, "1808": {"address": "Kapele", "lat": 45.9310499, "lon": 15.6780936, "uuid": "S1"}, "56284": {"address": "hrvaska4", "lat": 45.8121115, "lon": 15.816119, "uuid": "H4"}, "60310": {"address": "hrvaska9", "lat": 45.882431, "lon": 15.7389888, "uuid": "H9"}, "79013": {"address": "Cerklje ob Krki", "lat": 45.8844935, "lon": 15.5231733, "uuid": "S5"}, "84418": {"address": "Ulica 11. novembra", "lat": 45.9419906, "lon": 15.4781109, "uuid": "S6"}, "92232": {"address": "Ulica stare pravde", "lat": 45.9046721, "lon": 15.5934382, "uuid": "S8"}, "89437": {"address": "Kriska vas", "lat": 45.8913565, "lon": 15.5705945, "uuid": "S7"}, "95157": {"address": "Ulica bratov Gerjovicev", "lat": 45.8949345, "lon": 15.6602321, "uuid": "S9"}}, "edge": [[89437, 92232, 3418], [89437, 79013, 3860], [89437, 84418, 6138], [89437, 25590, 3394], [86774, 47460, 3456], [86774, 56284, 5273], [86774, 17940, 4945], [63502, 31310, 3550], [63502, 60310, 3640], [63502, 95157, 3360], [79013, 89437, 3860], [79013, 84418, 5737], [18440, 60310, 3352], [18440, 31310, 7202], [92232, 89437, 3418], [92232, 25590, 3564], [92232, 95157, 3747], [92232, 2499, 4077], [60310, 63502, 3640], [60310, 31310, 3674], [60310, 18440, 3352], [47460, 17940, 3672], [47460, 86774, 3456], [47460, 56284, 9994], [95157, 63502, 3360], [95157, 92232, 3747], [95157, 1808, 4002], [95157, 31310, 4137], [95157, 2499, 5620], [31310, 95157, 4137], [31310, 63502, 3550], [31310, 60310, 3674], [31310, 1808, 4002], [31310, 18440, 7202], [17940, 25590, 4998], [17940, 47460, 3672], [17940, 56284, 10457], [17940, 86774, 4945], [56284, 86774, 5273], [56284, 17940, 10457], [56284, 47460, 9994], [2499, 1808, 4011], [2499, 92232, 4077], [2499, 95157, 5620], [25590, 92232, 3564], [25590, 89437, 3394], [25590, 17940, 4998], [1808, 2499, 4011], [1808, 95157, 4002], [1808, 31310, 4002], [84418, 89437, 6138], [84418, 79013, 5737]]}
+{
+  "nodes": {
+    "2499": {
+      "address": "Globoko",
+      "lat": 45.9540673,
+      "lon": 15.6339655,
+      "uuid": "S2"
+    },
+    "63502": {
+      "address": "hrvaska5",
+      "lat": 45.8936019,
+      "lon": 15.6962282,
+      "uuid": "H5"
+    },
+    "47460": {
+      "address": "hrvaska1",
+      "lat": 45.839665,
+      "lon": 15.6863823,
+      "uuid": "H1"
+    },
+    "17940": {
+      "address": "Jesenice",
+      "lat": 45.8595037,
+      "lon": 15.6897957,
+      "uuid": "S3"
+    },
+    "18440": {
+      "address": "hrvaska6",
+      "lat": 45.8606795,
+      "lon": 15.8031616,
+      "uuid": "H6"
+    },
+    "86774": {
+      "address": "hrvaska2",
+      "lat": 45.8031663,
+      "lon": 15.7101405,
+      "uuid": "H2"
+    },
+    "25590": {
+      "address": "Topliska cesta",
+      "lat": 45.89337,
+      "lon": 15.6214561,
+      "uuid": "S4"
+    },
+    "31310": {
+      "address": "hrvaska3",
+      "lat": 45.9171602,
+      "lon": 15.7280016,
+      "uuid": "H3"
+    },
+    "1808": {
+      "address": "Kapele",
+      "lat": 45.9310499,
+      "lon": 15.6780936,
+      "uuid": "S1"
+    },
+    "56284": {
+      "address": "hrvaska4",
+      "lat": 45.8121115,
+      "lon": 15.816119,
+      "uuid": "H4"
+    },
+    "60310": {
+      "address": "hrvaska9",
+      "lat": 45.882431,
+      "lon": 15.7389888,
+      "uuid": "H9"
+    },
+    "79013": {
+      "address": "Cerklje ob Krki",
+      "lat": 45.8844935,
+      "lon": 15.5231733,
+      "uuid": "S5"
+    },
+    "84418": {
+      "address": "Ulica 11. novembra",
+      "lat": 45.9419906,
+      "lon": 15.4781109,
+      "uuid": "S6"
+    },
+    "92232": {
+      "address": "Ulica stare pravde",
+      "lat": 45.9046721,
+      "lon": 15.5934382,
+      "uuid": "S8"
+    },
+    "89437": {
+      "address": "Kriska vas",
+      "lat": 45.8913565,
+      "lon": 15.5705945,
+      "uuid": "S7"
+    },
+    "95157": {
+      "address": "Ulica bratov Gerjovicev",
+      "lat": 45.8949345,
+      "lon": 15.6602321,
+      "uuid": "S9"
+    }
+  },
+  "edge": [
+    [
+      89437,
+      92232,
+      3418
+    ],
+    [
+      89437,
+      79013,
+      3860
+    ],
+    [
+      89437,
+      84418,
+      6138
+    ],
+    [
+      89437,
+      25590,
+      3394
+    ],
+    [
+      86774,
+      47460,
+      3456
+    ],
+    [
+      86774,
+      56284,
+      5273
+    ],
+    [
+      86774,
+      17940,
+      4945
+    ],
+    [
+      63502,
+      31310,
+      3550
+    ],
+    [
+      63502,
+      60310,
+      3640
+    ],
+    [
+      63502,
+      95157,
+      3360
+    ],
+    [
+      79013,
+      89437,
+      3860
+    ],
+    [
+      79013,
+      84418,
+      5737
+    ],
+    [
+      18440,
+      60310,
+      3352
+    ],
+    [
+      18440,
+      31310,
+      7202
+    ],
+    [
+      92232,
+      89437,
+      3418
+    ],
+    [
+      92232,
+      25590,
+      3564
+    ],
+    [
+      92232,
+      95157,
+      3747
+    ],
+    [
+      92232,
+      2499,
+      4077
+    ],
+    [
+      60310,
+      63502,
+      3640
+    ],
+    [
+      60310,
+      31310,
+      3674
+    ],
+    [
+      60310,
+      18440,
+      3352
+    ],
+    [
+      47460,
+      17940,
+      3672
+    ],
+    [
+      47460,
+      86774,
+      3456
+    ],
+    [
+      47460,
+      56284,
+      9994
+    ],
+    [
+      95157,
+      63502,
+      3360
+    ],
+    [
+      95157,
+      92232,
+      3747
+    ],
+    [
+      95157,
+      1808,
+      4002
+    ],
+    [
+      95157,
+      31310,
+      4137
+    ],
+    [
+      95157,
+      2499,
+      5620
+    ],
+    [
+      31310,
+      95157,
+      4137
+    ],
+    [
+      31310,
+      63502,
+      3550
+    ],
+    [
+      31310,
+      60310,
+      3674
+    ],
+    [
+      31310,
+      1808,
+      4002
+    ],
+    [
+      31310,
+      18440,
+      7202
+    ],
+    [
+      17940,
+      25590,
+      4998
+    ],
+    [
+      17940,
+      47460,
+      3672
+    ],
+    [
+      17940,
+      56284,
+      10457
+    ],
+    [
+      17940,
+      86774,
+      4945
+    ],
+    [
+      56284,
+      86774,
+      5273
+    ],
+    [
+      56284,
+      17940,
+      10457
+    ],
+    [
+      56284,
+      47460,
+      9994
+    ],
+    [
+      2499,
+      1808,
+      4011
+    ],
+    [
+      2499,
+      92232,
+      4077
+    ],
+    [
+      2499,
+      95157,
+      5620
+    ],
+    [
+      25590,
+      92232,
+      3564
+    ],
+    [
+      25590,
+      89437,
+      3394
+    ],
+    [
+      25590,
+      17940,
+      4998
+    ],
+    [
+      1808,
+      2499,
+      4011
+    ],
+    [
+      1808,
+      95157,
+      4002
+    ],
+    [
+      1808,
+      31310,
+      4002
+    ],
+    [
+      84418,
+      89437,
+      6138
+    ],
+    [
+      84418,
+      79013,
+      5737
+    ]
+  ]
+}
