+++ src/modules/create_graph/data_parser/parse_posts.py
@@ -1,14 +1,14 @@
-
 import csv
+from ...utils.structures.post import Post
 
-from ..pojo.post import Post
 
 class PostHandler:
 
     def __init__(self):
         self.posts = []
 
-    def is_number(self, str):
+    @staticmethod
+    def is_number(str):
         try:
             return float(str)
         except ValueError:
@@ -20,7 +20,7 @@ class PostHandler:
             csv_reader = csv.reader(csv_file, delimiter=',')
 
             for row in csv_reader:
-                #address = [row[13], row[14], row[15]]
+                # address = [row[13], row[14], row[15]]
                 address = [row[0]]
                 uuid = [row[1]]
                 if (self.is_number(row[2]) != None and self.is_number(row[3]) != None):
@@ -41,8 +41,11 @@ class PostHandler:
         return posts
 
     def read_postal_offices(self, post_paths):
-        ''' Postal offices are read from csv file and than added to array
-        '''
+        """
+        Postal offices are read from csv file and than added to array
+        :param post_paths:
+        :return:
+        """
         self.posts = []
         for name, post_path in post_paths.items():
             if 'si' in name:
