+++ src/modules/utils/structures/parcel.py
@@ -4,7 +4,6 @@ class Parcel:
     Target is a destination Node
     Default package volume for poc is 1, for easier UUID unpacking
     """
-
     def __init__(self, uuid, target, volume):
         self.volume = volume
         self.target = target
