+++ src/modules/utils/structures/plan.py
@@ -4,7 +4,6 @@ class Plan:
     vehicles in this partition, all deliveries in this partition and
     the partition object, which is a GraphProcessor object
     """
-
     def __init__(self, vehicles, deliveries, deliveries_req, partition):
         self.vehicles = vehicles
         self.deliveries = deliveries
