+++ src/modules/utils/structures/vehicle.py
@@ -4,7 +4,6 @@ class Vehicle:
     name is usually uuid
     start_node is name/matching UUID
     """
-
     def __init__(self, name, start_node, parcels, capacity=200):
         self.capacity = capacity
         self.name = name
